chore: split shared `types`
