+++ src/shared/index.ts
@@ -1,1 +1,2 @@
-export * from './types.js'
+export * from './plugin.js'
+export * from './network.js'
