+++ src/shared/plugin.ts
@@ -1,56 +1,7 @@
 import type { QRCodeToDataURLOptions } from 'qrcode'
+import type { SocialShareNetwork } from './network'
 
 export type SocialShareNetworkData = Record<string, SocialShareNetwork>
-
-/**
- * @deprecated use {@link SocialShareNetworkWithName} instead
- */
-export type SocialShareNetworkItem = SocialShareNetworkWithName
-
-/**
- * Social share network with name
- */
-export type SocialShareNetworkWithName = SocialShareNetwork & {
-  /**
-   * Sharer name
-   */
-  name: string
-}
-
-/**
- * Social share network
- */
-export type SocialShareColor =
-  | string
-  | {
-      /**
-       * color for dark mode
-       */
-      dark: string
-      /**
-       * color for light mode
-       */
-      light: string
-    }
-export interface SocialShareNetwork {
-  icon: SocilaShareIcon
-  type: SocialShareType
-  color?: SocialShareColor
-  sharer?: string
-}
-export type SocialShareType = 'direct' | 'popup' | 'qrcode'
-export type SocilaShareIcon =
-  | string
-  | {
-      /**
-       * icon for dark mode
-       */
-      dark: string
-      /**
-       * icon for light mode
-       */
-      light: string
-    }
 
 /**
  * Front-matter
